refac(repo sync): extract into Handler
